refactor: create UserAvatar as canonical reusable avatar component
Replace AvatarWithBorder usage across all components with UserAvatar
so future avatar changes only need to update one component. Also fixes
badge detail page "earned by" section to show avatar borders.
