+++ src/components/ui/index.ts
@@ -4,6 +4,7 @@ export { default as UIBadge } from "./Badge";
 export { default as Input } from "./Input";
 export { default as Avatar } from "./Avatar";
 export { default as AvatarWithBorder } from "./AvatarWithBorder";
+export { default as UserAvatar } from "./UserAvatar";
 export { default as DateRangePicker } from "./DateRangePicker";
 export { default as OtpCodeInput } from "./OtpCodeInput";
 export {
